Moved all ASCII images to a new file
