+++ app/ascii-images.rb
@@ -1,0 +1,120 @@
+def display_city
+    pic = "
+    __    __                         ___      _
+    |  |  |  |       /|             |   |   _/ \_
+    |  |  |  |   _  | |__           |   |_-/     \-  _ _
+    |__|  |  |  |_| | | |  |/\_     |   |  \     /  |___|
+    |  |  |  |  | | __| |  |   |_   |   |   |___|   |   |
+    |  |  |^ |  | ||  | |  |   | |__|   |   |   |   |   |
+    |  |  |  |  | ||  | |  |   |/\  |   |   |   |   |   |
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~~/  \~~~~~~~~~~~~~~~~~~~~~~~
+    ~ ~~  ~ ~~ ~~~ ~ ~ ~~ ~~ ~~ \   \__   ~  ~  ~~~~ ~~~ ~~
+    ~~ ~ ~ ~~~ ~~  ~~ ~~~~~~~~~~ \   \o\  ~~ ~ ~~~~ ~ ~ ~~~
+    ~ ~~~~~~~~ ~ ~ ~~ ~ ~ ~ ~ ~~~ \   \o\=   ~~ ~~  ~~ ~ ~~
+    ~ ~ ~ ~~~~~~~ ~  ~~ ~~ ~ ~~ ~ ~ ~~ ~ ~ ~~ ~~~ ~ ~ ~ ~ ~~~~
+    "
+    puts pic
+end
+
+def banner 
+    banner = "       
+                  000000101101010101010101010101101011000101101010101011000101010101010101010101001010101010110101010101010101010101010101001010110101        
+                  010010                             ~+? ?~ ~?~?++~?~~~               ~~        +                                               010101          
+                  101010                    ~~  :~:~+ ~  ?++++++++++,        ~+         ~:      ~:?:~        ~                                  010010            
+                  100101                 :~,++~ ~ ?,++.     ?+++++~                   ++   +   +++++++++++?~ ~++~       ~                       101010              
+                  010110     ?+++++++++?+?+?+?+:+~+ ~ ?+~  ,+++              ~++++++  ++++++++?++++++++++++++++++++++++++++~~                   101001                
+                  010110  +++++++?++++++?~?++++    ~+~:    ++              ~?+~+++++++++++++++++++++++++++++++++++++++++++++                    101001                  
+                  101011 :: ~  ~? +++++++++++~    ++,:~                 +  ? + ~??+++++++++++++++++++++++++++++++++     +~                      011001                   
+                  010101        ~?++++++++?+++++~+++++?               :~+   : +~?+~?++++++++++++ +++++++++~?+++++++      +~                     101010         
+                  101011        :++++++++++++++++++??+ ~               ~:+++? ++,++?++++ ?  +++++++:++~++++++++++++++ ~   ~                     111000           
+                  101010        :++++++++++  ++++?++~  ~                ?++ ? +++?~+~?+?~ ++,++++~,+++ ++++++++:+++? + ~                        101001             
+                  011001       ?+++++++++++:+ ~++                     ++~?  +~,+++      :~++?+ + ~~:+++++++++ +++++?~  ++                       000000               
+                  010110       ++++++++++++++++:                     ~+++~ ~  ~~,:++++++~ ~++++ + ?++++++++++++++  ?+   +                       110100               
+                  101010       +++++++++++++++~                       +:?+?,     : +~+++~+++++?++?? +++++++++++++++ ?  ++                       101010            
+                  010101        ,   ++++++++                         ?+~+++?++ ~+: ~~~++++++++++~++++++++++++++++++   :                         001010       
+                  010101        +?++++     +                      ~  :++++++++++++++~++++? ++++++ +++++ ??+~+++++++                             100010         
+                  010110          +++:       ~                     +?+++++++ ?+++++++ +++++     ~+++++++++?~++++++~+                            100101           
+                  011010           ++?  +   ~                     ~?+++++ +++++++~++++ ++++?+~     ++++:  ++ +?                                 101001             
+                  010101            ~??+~                         :?++ ?+++++++++++++++~+~+:       ++++    ~?+ ~    ?                           110101               
+                  010110                ~+    ~                   , ?+?+++++++?+ +++++++~           ++       ~+~    ~:                          110010                 
+                  010101                  :~ ???+?                 , ~+~ ++++?+~+?+++++?~++         ?+      ?      ~ ~~                         101010                    
+                  001011                    ++++++ ~:                +:    ++++~+  +++++?+                 ?++    ~?~                           101011                
+                  011010                   ~+++? ?++++                      ~~+++++ ++++~                   :+, ?:??  ~+                        010101           
+                  010101                   + + +++++??++~~                   :++++?:: ?~                     ?+~ +++? ~  ~+~  ~~                101010             
+                  010101                   ++++++++++++++++?                 +~?+++++++                        +    ~    : +?+~                 101001               
+                  110001                    ++ +~++++++++++~                 ?++:?++~?+                              ~        +:                101010                 
+                  010101                     ++ +?~+++++++                   +++++++ ++   +                           ~ +?                  ~   000000                   
+                  100110                       ++++~++++++                   +~+? +???  ?+                           +++++??+           ~       000100                     
+                  010101                       ?++?+?++++?                   ~++++~++   ++                        ++++++++++++                  001001              
+                  010101                       ++++++?+                       +:++++~   +~                       ++++++++++++++                 100010         
+                  111010                       ++++ ██╗    ██╗ █████╗ ██████╗      ██████╗  █████╗ ███╗   ███╗███████╗███████╗                  101001           
+                  000101                       +++  ██║    ██║██╔══██╗██╔══██╗    ██╔════╝ ██╔══██╗████╗ ████║██╔════╝██╔════╝                  101001             
+                  101010                       ++   ██║ █╗ ██║███████║██████╔╝    ██║  ███╗███████║██╔████╔██║█████╗  ███████╗                  110010               
+                  101010                       +++  ██║███╗██║██╔══██║██╔══██╗    ██║   ██║██╔══██║██║╚██╔╝██║██╔══╝  ╚════██║                  101000                 
+                  101011                       ++   ╚███╔███╔╝██║  ██║██║  ██║    ╚██████╔╝██║  ██║██║ ╚═╝ ██║███████╗███████║                  001000                   
+                  101010                             ╚══╝╚══╝ ╚═╝  ╚═╝╚═╝  ╚═╝     ╚═════╝ ╚═╝  ╚═╝╚═╝     ╚═╝╚══════╝╚══════╝                  010111                     
+                  000000101101010101010101010101101011000101101010101011000101010101010101010101001010101010110101010101010101010101010101001010110101                
+"                    
+    puts banner 
+end
+
+def winningmove
+    move = "                           
+
+
+
+
+
+
+
+                        ██╗███╗   ██╗    ███╗   ██╗██╗   ██╗ ██████╗██╗     ███████╗ █████╗ ██████╗     ██╗    ██╗ █████╗ ██████╗                                         
+                        ██║████╗  ██║    ████╗  ██║██║   ██║██╔════╝██║     ██╔════╝██╔══██╗██╔══██╗    ██║    ██║██╔══██╗██╔══██╗                                        
+                        ██║██╔██╗ ██║    ██╔██╗ ██║██║   ██║██║     ██║     █████╗  ███████║██████╔╝    ██║ █╗ ██║███████║██████╔╝                                        
+                        ██║██║╚██╗██║    ██║╚██╗██║██║   ██║██║     ██║     ██╔══╝  ██╔══██║██╔══██╗    ██║███╗██║██╔══██║██╔══██╗                                        
+                        ██║██║ ╚████║    ██║ ╚████║╚██████╔╝╚██████╗███████╗███████╗██║  ██║██║  ██║    ╚███╔███╔╝██║  ██║██║  ██║                                        
+                        ╚═╝╚═╝  ╚═══╝    ╚═╝  ╚═══╝ ╚═════╝  ╚═════╝╚══════╝╚══════╝╚═╝  ╚═╝╚═╝  ╚═╝     ╚══╝╚══╝ ╚═╝  ╚═╝╚═╝  ╚═╝                                        
+                        ████████╗██╗  ██╗███████╗     ██████╗ ███╗   ██╗██╗  ██╗   ██╗    ██╗    ██╗ █████╗ ██╗   ██╗    ████████╗ ██████╗     ██╗    ██╗██╗███╗   ██╗    
+                        ╚══██╔══╝██║  ██║██╔════╝    ██╔═══██╗████╗  ██║██║  ╚██╗ ██╔╝    ██║    ██║██╔══██╗╚██╗ ██╔╝    ╚══██╔══╝██╔═══██╗    ██║    ██║██║████╗  ██║    
+                           ██║   ███████║█████╗      ██║   ██║██╔██╗ ██║██║   ╚████╔╝     ██║ █╗ ██║███████║ ╚████╔╝        ██║   ██║   ██║    ██║ █╗ ██║██║██╔██╗ ██║    
+                           ██║   ██╔══██║██╔══╝      ██║   ██║██║╚██╗██║██║    ╚██╔╝      ██║███╗██║██╔══██║  ╚██╔╝         ██║   ██║   ██║    ██║███╗██║██║██║╚██╗██║    
+                           ██║   ██║  ██║███████╗    ╚██████╔╝██║ ╚████║███████╗██║       ╚███╔███╔╝██║  ██║   ██║          ██║   ╚██████╔╝    ╚███╔███╔╝██║██║ ╚████║    
+                           ╚═╝   ╚═╝  ╚═╝╚══════╝     ╚═════╝ ╚═╝  ╚═══╝╚══════╝╚═╝        ╚══╝╚══╝ ╚═╝  ╚═╝   ╚═╝          ╚═╝    ╚═════╝      ╚══╝╚══╝ ╚═╝╚═╝  ╚═══╝    
+                        ██╗███████╗    ████████╗ ██████╗     ███╗   ██╗ ██████╗ ████████╗    ██████╗ ██╗      █████╗ ██╗   ██╗                                            
+                        ██║██╔════╝    ╚══██╔══╝██╔═══██╗    ████╗  ██║██╔═══██╗╚══██╔══╝    ██╔══██╗██║     ██╔══██╗╚██╗ ██╔╝                                            
+                        ██║███████╗       ██║   ██║   ██║    ██╔██╗ ██║██║   ██║   ██║       ██████╔╝██║     ███████║ ╚████╔╝                                             
+                        ██║╚════██║       ██║   ██║   ██║    ██║╚██╗██║██║   ██║   ██║       ██╔═══╝ ██║     ██╔══██║  ╚██╔╝                                              
+                        ██║███████║       ██║   ╚██████╔╝    ██║ ╚████║╚██████╔╝   ██║       ██║     ███████╗██║  ██║   ██║██╗██╗██╗                                      
+                        ╚═╝╚══════╝       ╚═╝    ╚═════╝     ╚═╝  ╚═══╝ ╚═════╝    ╚═╝       ╚═╝     ╚══════╝╚═╝  ╚═╝   ╚═╝╚═╝╚═╝╚═╝
+
+
+
+"
+puts move 
+end
+
+def won 
+    won = "
+
+
+
+                                                        ██╗   ██╗ ██████╗ ██╗   ██╗    ██╗    ██╗ ██████╗ ███╗   ██╗██╗
+                                                        ╚██╗ ██╔╝██╔═══██╗██║   ██║    ██║    ██║██╔═══██╗████╗  ██║██║
+                                                         ╚████╔╝ ██║   ██║██║   ██║    ██║ █╗ ██║██║   ██║██╔██╗ ██║██║
+                                                          ╚██╔╝  ██║   ██║██║   ██║    ██║███╗██║██║   ██║██║╚██╗██║╚═╝
+                                                           ██║   ╚██████╔╝╚██████╔╝    ╚███╔███╔╝╚██████╔╝██║ ╚████║██╗
+                                                           ╚═╝    ╚═════╝  ╚═════╝      ╚══╝╚══╝  ╚═════╝ ╚═╝  ╚═══╝╚═╝
+                                                                                                                "    
+puts won
+end
+
+def overhead
+    overhead = "
+    ██████╗  █████╗ ███╗   ███╗███████╗     ██████╗ ██╗   ██╗███████╗██████╗ 
+    ██╔════╝ ██╔══██╗████╗ ████║██╔════╝    ██╔═══██╗██║   ██║██╔════╝██╔══██╗
+    ██║  ███╗███████║██╔████╔██║█████╗      ██║   ██║██║   ██║█████╗  ██████╔╝
+    ██║   ██║██╔══██║██║╚██╔╝██║██╔══╝      ██║   ██║╚██╗ ██╔╝██╔══╝  ██╔══██╗
+    ╚██████╔╝██║  ██║██║ ╚═╝ ██║███████╗    ╚██████╔╝ ╚████╔╝ ███████╗██║  ██║
+    ╚═════╝ ╚═╝  ╚═╝╚═╝     ╚═╝╚══════╝     ╚═════╝   ╚═══╝  ╚══════╝╚═╝  ╚═╝
+                  "
+    puts overhead
+end
